Merge integrate-ai-emotions into FEATURE-AI-Agent (#17)
* AI_Basic_Movement

Navigation & NavMesh appears to be setup

To implement
- AI Emotions for the KBA

* Updates

Still not 100% happy with AI collision but now it pauses, and redirects if it collides with something which is a plus

* AI Can now collide with player

* Emotion Emojis Added to AI

Emotion Emojis Added to AI

* Updated base movement

* updates

* Updates

Separation of AI_Movement.gd into AI_Movement.gd and EmojiManager.gd

* Affordances

Jumping and Light up

* AI Affordance update

Cube turns red when AI approaches and angry. White when AI is in any other mood.

* Updates to existing code.

* Updates to affordance cube

* Model for AI

Model does load into scene however,

1) cant figure out how to have animations work.

2) For some unknown reason the AI refuses to be on the ground and chooses to fly everywhere ?

* Model Position & loading

* AI_Movement with model

* started work on final scene

* added teacher agent with simple goals

* added log window

* logger updates

* added fuzzy triangle for emotions

* code cleanup

* added TeacherAgent prefab

---------

Co-authored-by: Jinxed Nation William Halling <32233703@student.murdoch.edu.au>